fix: renamed VehicleStopNearYou -> VehicleRentalStationNearYou
